data-pipelines/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-redis==3.5.1
apache-airflow-providers-docker==3.9.1
apache-airflow-providers-http==4.11.1
pandas==2.1.4
numpy==1.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
celery==5.3.4
flower==2.0.1
kubernetes==28.1.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
